Refactored py3 logic to asgi module for py2 compat
